ERR: Problem with stl, added typecast to solve following cough on Borland Compiler... Error E2015 D:\vtknew\IO\vtkImageReader.cxx 594: Ambiguity between 'std::fpos<std::mbstate_t>::operator -(int) const' and 'std::fpos<std::mbstate_t>::operator -(long) const' in function vtkImageReader::GetHeaderSize(int)

+++ IO/vtkImageReader.cxx
@@ -591,7 +591,7 @@ unsigned long vtkImageReader::GetHeaderSize(int idx)
     // Get the size of the header from the size of the image
     this->File->seekg(0,ios::end);
     
-    return (unsigned long)(this->File->tellg() - 
+    return (unsigned long)((unsigned long)this->File->tellg() - 
       (unsigned long)this->DataIncrements[this->GetFileDimensionality()]);
     }
